refactor(pip): extract reusable PiP lifecycle hooks
