+++ react/features/pip/hooks.ts
@@ -1,9 +1,12 @@
 import React, { useEffect, useRef } from 'react';
 
+import { IStore } from '../app/types';
 import IconUserSVG from '../base/icons/svg/user.svg?raw';
 import { IParticipant } from '../base/participants/types';
+import { ITrack } from '../base/tracks/types';
 import { TILE_ASPECT_RATIO } from '../filmstrip/constants';
 
+import { handlePiPLeaveEvent, handlePipEnterEvent, handleWindowBlur, handleWindowFocus } from './actions';
 import { renderAvatarOnCanvas } from './functions';
 import logger from './logger';
 
@@ -18,6 +21,12 @@ const CANVAS_HEIGHT = Math.floor(CANVAS_WIDTH / TILE_ASPECT_RATIO);
  * We manually request frames after drawing to ensure capture.
  */
 const CANVAS_FRAME_RATE = 0;
+
+/**
+ * Delay before exiting PiP on focus to avoid Chrome getting stuck when the PiP
+ * window is closed by browser chrome controls.
+ */
+const WINDOW_FOCUS_EXIT_DELAY_MS = 100;
 
 /**
  * Options for the useCanvasAvatar hook.
@@ -180,4 +189,162 @@ export function useCanvasAvatar(options: IUseCanvasAvatarOptions): IUseCanvasAva
     return {
         canvasStreamRef: streamRef
     };
+}
+
+/**
+ * Keeps the hidden PiP video element in sync with the currently selected video
+ * source, switching between the participant track and the generated avatar
+ * canvas stream.
+ *
+ * @param {React.RefObject<HTMLVideoElement | null>} videoRef - Ref to the PiP
+ * video element.
+ * @param {React.MutableRefObject<ITrack | undefined>} previousTrackRef - Ref to
+ * the previously attached video track.
+ * @param {ITrack | undefined} videoTrack - Current video track to display.
+ * @param {boolean} shouldShowAvatar - Whether the avatar canvas should be used
+ * instead of a participant video track.
+ * @param {React.MutableRefObject<MediaStream | null>} canvasStreamRef - Ref to
+ * the generated avatar canvas stream.
+ * @returns {void}
+ */
+export function usePiPVideoSource(
+        videoRef: React.RefObject<HTMLVideoElement | null>,
+        previousTrackRef: React.MutableRefObject<ITrack | undefined>,
+        videoTrack: ITrack | undefined,
+        shouldShowAvatar: boolean,
+        canvasStreamRef: React.MutableRefObject<MediaStream | null>
+) {
+    useEffect(() => {
+        const videoElement = videoRef.current;
+
+        if (!videoElement) {
+            return;
+        }
+
+        const previousTrack = previousTrackRef.current;
+
+        if (previousTrack?.jitsiTrack) {
+            try {
+                previousTrack.jitsiTrack.detach(videoElement);
+            } catch (error) {
+                logger.error('Error detaching previous track:', error);
+            }
+        }
+
+        if (shouldShowAvatar) {
+            const canvasStream = canvasStreamRef.current;
+
+            if (canvasStream && videoElement.srcObject !== canvasStream) {
+                videoElement.srcObject = canvasStream;
+            }
+        } else if (videoTrack?.jitsiTrack) {
+            videoTrack.jitsiTrack.attach(videoElement)
+                .catch((error: Error) => {
+                    logger.error('Error attaching video track:', error);
+                });
+        }
+
+        previousTrackRef.current = videoTrack;
+
+        return () => {
+            if (videoTrack?.jitsiTrack && videoElement) {
+                try {
+                    videoTrack.jitsiTrack.detach(videoElement);
+                } catch (error) {
+                    logger.error('Error during cleanup:', error);
+                }
+            }
+        };
+    }, [ canvasStreamRef, previousTrackRef, shouldShowAvatar, videoRef, videoTrack ]);
+}
+
+/**
+ * Registers the window and document events that automatically enter/exit PiP
+ * based on app focus.
+ *
+ * @param {Object} videoRef - Ref to the PiP video element.
+ * @param {Function} dispatch - Redux dispatch function.
+ * @returns {void}
+ */
+export function usePiPWindowLifecycle(
+        videoRef: React.RefObject<HTMLVideoElement | null>,
+        dispatch: IStore['dispatch']
+) {
+    useEffect(() => {
+        const videoElement = videoRef.current;
+
+        if (!videoElement) {
+            return;
+        }
+
+        const onWindowBlur = () => dispatch(handleWindowBlur(videoElement));
+        const onWindowFocus = () => {
+            setTimeout(() => {
+                dispatch(handleWindowFocus());
+            }, WINDOW_FOCUS_EXIT_DELAY_MS);
+        };
+        const onVisibilityChange = () => {
+            if (document.hidden) {
+                onWindowBlur();
+            }
+        };
+        const checkFocusAndEnterPiP = () => {
+            if (!document.hasFocus()) {
+                onWindowBlur();
+            }
+        };
+
+        window.addEventListener('blur', onWindowBlur);
+        window.addEventListener('focus', onWindowFocus);
+        document.addEventListener('visibilitychange', onVisibilityChange);
+
+        if (videoElement.readyState >= 1) {
+            checkFocusAndEnterPiP();
+        } else {
+            videoElement.addEventListener('loadedmetadata', checkFocusAndEnterPiP, { once: true });
+        }
+
+        return () => {
+            window.removeEventListener('blur', onWindowBlur);
+            window.removeEventListener('focus', onWindowFocus);
+            document.removeEventListener('visibilitychange', onVisibilityChange);
+            videoElement.removeEventListener('loadedmetadata', checkFocusAndEnterPiP);
+        };
+    }, [ dispatch, videoRef ]);
+}
+
+/**
+ * Registers browser PiP enter/leave events for the hidden PiP video element
+ * and forwards them to the Redux handlers.
+ *
+ * @param {Object} videoRef - Ref to the PiP video element.
+ * @param {Function} dispatch - Redux dispatch function.
+ * @returns {void}
+ */
+export function usePiPBrowserEvents(
+        videoRef: React.RefObject<HTMLVideoElement | null>,
+        dispatch: IStore['dispatch']
+) {
+    useEffect(() => {
+        const videoElement = videoRef.current;
+
+        if (!videoElement) {
+            return;
+        }
+
+        const onEnterPiP = () => {
+            dispatch(handlePipEnterEvent());
+        };
+        const onLeavePiP = () => {
+            dispatch(handlePiPLeaveEvent());
+        };
+
+        videoElement.addEventListener('enterpictureinpicture', onEnterPiP);
+        videoElement.addEventListener('leavepictureinpicture', onLeavePiP);
+
+        return () => {
+            videoElement.removeEventListener('enterpictureinpicture', onEnterPiP);
+            videoElement.removeEventListener('leavepictureinpicture', onLeavePiP);
+        };
+    }, [ dispatch, videoRef ]);
 }
